generalize

+++ crates/libs/metadata/src/array_info.rs
@@ -1,5 +1,5 @@
 #[derive(Copy, Clone)]
 pub enum ArrayInfo {
-    Relative(i16),
-    Fixed(i32),
+    Relative(usize),
+    Fixed(usize),
 }
